pyproject.toml
[tool.poetry]
name = "ansible-cisco-radkit"
version = "2.0.0"
description = "Ansible Galaxy Collection for Cisco RADKit"
authors = ["Scott Dozier <scdozier@cisco.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">= 3.9, < 3.12"
ansible = "^8.7.0"
pproxy = "^2.7.0"
httpx = "^0.27.2"
cisco-radkit-client = "^1.8.5"
cisco-radkit-genie = "^1.8.5"

[tool.poetry.group.dev.dependencies]
ansible-lint = "^6.0.0"
black = "^23.0.0"
isort = "^5.12.0"
yamllint = "^1.32.0"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.pytest_cache
  | \.venv
  | build
  | dist
  | tmp
)/
'''

